Refactor Barsukas string helpers and add accessor tests
